checkpoint got stdout/stderr data packets working with new remote handler

+++ pkg/shexec/shexec.go
@@ -12,6 +12,7 @@ import (
 	"os"
 	"os/exec"
 	"strings"
+	"sync"
 	"syscall"
 	"time"
 
@@ -27,14 +28,48 @@ const MaxRows = 1024
 const MaxCols = 1024
 
 type ShExecType struct {
-	FileNames *base.CommandFileNames
-	Cmd       *exec.Cmd
-	CmdPty    *os.File
-	StartTs   time.Time
+	StartTs         time.Time
+	RunPacket       *packet.RunPacketType
+	FileNames       *base.CommandFileNames
+	Cmd             *exec.Cmd
+	CmdPty          *os.File
+	FdReaders       map[int]*os.File
+	FdWriters       map[int]*os.File
+	CloseAfterStart []*os.File
+}
+
+func MakeShExec(pk *packet.RunPacketType) *ShExecType {
+	return &ShExecType{
+		StartTs:   time.Now(),
+		RunPacket: pk,
+		FdReaders: make(map[int]*os.File),
+		FdWriters: make(map[int]*os.File),
+	}
 }
 
 func (c *ShExecType) Close() {
-	c.CmdPty.Close()
+	if c.CmdPty != nil {
+		c.CmdPty.Close()
+	}
+	for _, fd := range c.FdReaders {
+		fd.Close()
+	}
+	for _, fd := range c.FdWriters {
+		fd.Close()
+	}
+	for _, fd := range c.CloseAfterStart {
+		fd.Close()
+	}
+}
+
+func (c *ShExecType) MakeCmdStartPacket() *packet.CmdStartPacketType {
+	startPacket := packet.MakeCmdStartPacket()
+	startPacket.Ts = time.Now().UnixMilli()
+	startPacket.SessionId = c.RunPacket.SessionId
+	startPacket.CmdId = c.RunPacket.CmdId
+	startPacket.Pid = c.Cmd.Process.Pid
+	startPacket.MShellPid = os.Getpid()
+	return startPacket
 }
 
 func getEnvStrKey(envStr string) string {
@@ -93,7 +128,10 @@ func MakeExecCmd(pk *packet.RunPacketType, cmdTty *os.File) *exec.Cmd {
 }
 
 func MakeRunnerExec(cmdId string) (*exec.Cmd, error) {
-	msPath := base.GetMShellPath()
+	msPath, err := base.GetMShellPath()
+	if err != nil {
+		return nil, err
+	}
 	ecmd := exec.Command(msPath, cmdId)
 	return ecmd, nil
 }
@@ -124,19 +162,21 @@ func ValidateRunPacket(pk *packet.RunPacketType) error {
 	if pk.Type != packet.RunPacketStr {
 		return fmt.Errorf("run packet has wrong type: %s", pk.Type)
 	}
-	if pk.SessionId == "" {
-		return fmt.Errorf("run packet does not have sessionid")
-	}
-	_, err := uuid.Parse(pk.SessionId)
-	if err != nil {
-		return fmt.Errorf("invalid sessionid '%s' for command", pk.SessionId)
-	}
-	if pk.CmdId == "" {
-		return fmt.Errorf("run packet does not have cmdid")
-	}
-	_, err = uuid.Parse(pk.CmdId)
-	if err != nil {
-		return fmt.Errorf("invalid cmdid '%s' for command", pk.CmdId)
+	if pk.Detached {
+		if pk.SessionId == "" {
+			return fmt.Errorf("run packet does not have sessionid")
+		}
+		_, err := uuid.Parse(pk.SessionId)
+		if err != nil {
+			return fmt.Errorf("invalid sessionid '%s' for command", pk.SessionId)
+		}
+		if pk.CmdId == "" {
+			return fmt.Errorf("run packet does not have cmdid")
+		}
+		_, err = uuid.Parse(pk.CmdId)
+		if err != nil {
+			return fmt.Errorf("invalid cmdid '%s' for command", pk.CmdId)
+		}
 	}
 	if pk.Cwd != "" {
 		dirInfo, err := os.Stat(pk.Cwd)
@@ -164,13 +204,120 @@ func GetWinsize(p *packet.RunPacketType) *pty.Winsize {
 
 // when err is nil, the command will have already been started
 func RunCommand(pk *packet.RunPacketType, sender *packet.PacketSender) (*ShExecType, error) {
-	if pk.CmdId == "" {
-		pk.CmdId = uuid.New().String()
-	}
 	err := ValidateRunPacket(pk)
 	if err != nil {
 		return nil, err
 	}
+	if !pk.Detached {
+		return runCommandSimple(pk, sender)
+	} else {
+		return runCommandDetached(pk, sender)
+	}
+}
+
+// returns the *writer* to connect to process, reader is put in FdReaders
+func (cmd *ShExecType) makeReaderPipe(fdNum int) (*os.File, error) {
+	pr, pw, err := os.Pipe()
+	if err != nil {
+		return nil, err
+	}
+	cmd.FdReaders[fdNum] = pr
+	cmd.CloseAfterStart = append(cmd.CloseAfterStart, pw)
+	return pw, nil
+}
+
+// returns the *reader* to connect to process, writer is put in FdWriters
+func (cmd *ShExecType) makeWriterPipe(fdNum int) (*os.File, error) {
+	pr, pw, err := os.Pipe()
+	if err != nil {
+		return nil, err
+	}
+	cmd.FdWriters[fdNum] = pw
+	cmd.CloseAfterStart = append(cmd.CloseAfterStart, pr)
+	return pr, nil
+}
+
+func (cmd *ShExecType) MakeDataPacket(fdNum int, data []byte) *packet.DataPacketType {
+	pk := packet.MakeDataPacket()
+	pk.SessionId = cmd.RunPacket.SessionId
+	pk.CmdId = cmd.RunPacket.CmdId
+	pk.FdNum = fdNum
+	pk.Data = string(data)
+	return pk
+}
+
+func (cmd *ShExecType) runReadLoop(wg *sync.WaitGroup, fdNum int, fd *os.File, sender *packet.PacketSender) {
+	go func() {
+		defer fd.Close()
+		defer wg.Done()
+		buf := make([]byte, 4096)
+		for {
+			nr, err := fd.Read(buf)
+			pk := cmd.MakeDataPacket(fdNum, buf[0:nr])
+			if err == io.EOF {
+				pk.Eof = true
+				sender.SendPacket(pk)
+				break
+			} else if err != nil {
+				pk.Error = err.Error()
+				sender.SendPacket(pk)
+				break
+			} else {
+				sender.SendPacket(pk)
+			}
+		}
+	}()
+}
+
+func (cmd *ShExecType) RunIOAndWait(sender *packet.PacketSender) {
+	var wg sync.WaitGroup
+	wg.Add(len(cmd.FdReaders))
+	go func() {
+		for fdNum, fd := range cmd.FdReaders {
+			cmd.runReadLoop(&wg, fdNum, fd, sender)
+		}
+	}()
+	donePacket := cmd.WaitForCommand()
+	wg.Wait()
+	sender.SendPacket(donePacket)
+}
+
+func runCommandSimple(pk *packet.RunPacketType, sender *packet.PacketSender) (*ShExecType, error) {
+	cmd := MakeShExec(pk)
+	cmd.Cmd = exec.Command("bash", "-c", pk.Command)
+	UpdateCmdEnv(cmd.Cmd, pk.Env)
+	if pk.Cwd != "" {
+		cmd.Cmd.Dir = pk.Cwd
+	}
+	var err error
+	cmd.Cmd.Stdin, err = cmd.makeWriterPipe(0)
+	if err != nil {
+		cmd.Close()
+		return nil, err
+	}
+	cmd.Cmd.Stdout, err = cmd.makeReaderPipe(1)
+	if err != nil {
+		cmd.Close()
+		return nil, err
+	}
+	cmd.Cmd.Stderr, err = cmd.makeReaderPipe(2)
+	if err != nil {
+		cmd.Close()
+		return nil, err
+	}
+	err = cmd.Cmd.Start()
+	if err != nil {
+		cmd.Close()
+		return nil, err
+	}
+	for _, fd := range cmd.CloseAfterStart {
+		fd.Close()
+	}
+	cmd.CloseAfterStart = nil
+	return cmd, nil
+}
+
+func runCommandDetached(pk *packet.RunPacketType, sender *packet.PacketSender) (*ShExecType, error) {
 	fileNames, err := base.GetCommandFileNames(pk.SessionId, pk.CmdId)
 	if err != nil {
 		return nil, err
@@ -190,7 +337,7 @@ func RunCommand(pk *packet.RunPacketType, sender *packet.PacketSender) (*ShExecT
 	defer func() {
 		cmdTty.Close()
 	}()
-	startTs := time.Now()
+	rtn := MakeShExec(pk)
 	ecmd := MakeExecCmd(pk, cmdTty)
 	err = ecmd.Start()
 	if err != nil {
@@ -214,12 +361,10 @@ func RunCommand(pk *packet.RunPacketType, sender *packet.PacketSender) (*ShExecT
 			sender.SendErrorPacket(fmt.Sprintf("reading from stdin fifo: %v", copyFifoErr))
 		}
 	}()
-	return &ShExecType{
-		FileNames: fileNames,
-		Cmd:       ecmd,
-		CmdPty:    cmdPty,
-		StartTs:   startTs,
-	}, nil
+	rtn.FileNames = fileNames
+	rtn.Cmd = ecmd
+	rtn.CmdPty = cmdPty
+	return rtn, nil
 }
 
 func GetExitCode(err error) int {
@@ -233,16 +378,19 @@ func GetExitCode(err error) int {
 	}
 }
 
-func (c *ShExecType) WaitForCommand(cmdId string) *packet.CmdDonePacketType {
+func (c *ShExecType) WaitForCommand() *packet.CmdDonePacketType {
 	exitErr := c.Cmd.Wait()
 	endTs := time.Now()
 	cmdDuration := endTs.Sub(c.StartTs)
 	exitCode := GetExitCode(exitErr)
 	donePacket := packet.MakeCmdDonePacket()
 	donePacket.Ts = endTs.UnixMilli()
-	donePacket.CmdId = cmdId
+	donePacket.SessionId = c.RunPacket.SessionId
+	donePacket.CmdId = c.RunPacket.CmdId
 	donePacket.ExitCode = exitCode
 	donePacket.DurationMs = int64(cmdDuration / time.Millisecond)
-	os.Remove(c.FileNames.StdinFifo) // best effort (no need to check error)
+	if c.FileNames != nil {
+		os.Remove(c.FileNames.StdinFifo) // best effort (no need to check error)
+	}
 	return donePacket
 }
